requirements.txt
nltk
rouge-score
sacrebleu
torch
torchvision
torchaudio
pytorch-fid
pystoi
soundfile
pyyaml

